Rename BaseRuntime files to plugin-runtime.
