+++ sourcepawn/jit/plugin-runtime.h
