+++ sourcepawn/jit/x86/jit_x86.h
@@ -23,7 +23,7 @@
 #include <macro-assembler-x86.h>
 #include <am-vector.h>
 #include "jit_shared.h"
-#include "BaseRuntime.h"
+#include "plugin-runtime.h"
 #include "sp_vm_basecontext.h"
 #include "compiled-function.h"
 #include "opcodes.h"
